Replaces longs with ints
The long type no longer exists in Python 3.

Partially implements blueprint: nova-python3-mitaka

+++ nova/tests/unit/virt/vmwareapi/test_driver_api.py
@@ -399,7 +399,7 @@ class VMwareAPIVMTestCase(test.NoDBTestCase):
         vm = self._get_vm_record()
 
         # Check that m1.large above turned into the right thing.
-        mem_kib = long(self.type_data['memory_mb']) << 10
+        mem_kib = int(self.type_data['memory_mb']) << 10
         vcpus = self.type_data['vcpus']
         self.assertEqual(vm_info.max_mem_kb, mem_kib)
         self.assertEqual(vm_info.mem_kb, mem_kib)
@@ -439,7 +439,7 @@ class VMwareAPIVMTestCase(test.NoDBTestCase):
         """Check if the get_info returned values correspond to the instance
         object in the db.
         """
-        mem_kib = long(self.type_data['memory_mb']) << 10
+        mem_kib = int(self.type_data['memory_mb']) << 10
         self.assertEqual(info.state, pwr_state)
         self.assertEqual(info.max_mem_kb, mem_kib)
         self.assertEqual(info.mem_kb, mem_kib)
